+++ client/my-sites/customer-home/cards/constants.js
@@ -15,7 +15,6 @@ export const FEATURE_QUICK_START = 'home-feature-quick-start';
 export const FEATURE_STATS = 'home-feature-stats';
 export const FEATURE_SUPPORT = 'home-feature-support';
 export const NOTICE_SITE_LAUNCH_SELLER_UPSELL = 'home-site-launch-seller-upsell';
-export const NOTICE_SITE_RESUME_COPY = 'home-site-resume-copy';
 export const NOTICE_CELEBRATE_SITE_CREATION = 'home-notice-celebrate-site-creation';
 export const NOTICE_CELEBRATE_SITE_LAUNCH = 'home-notice-celebrate-site-launch';
 export const NOTICE_CELEBRATE_SITE_MIGRATION = 'home-notice-celebrate-site-migration';
@@ -37,6 +36,7 @@ export const TASK_RENEW_EXPIRING_PLAN = 'home-task-renew-expiring-plan';
 export const TASK_REACTIVATE_EXPIRED_PLAN = 'home-task-reactivate-expired-plan';
 export const TASK_REACTIVATE_ATOMIC_TRANSFER = 'home-task-reactivate-atomic-transfer';
 export const TASK_REACTIVATE_RESTORE_BACKUP = 'home-task-reactivate-restore-backup';
+export const TASK_SITE_RESUME_COPY = 'home-site-resume-copy';
 export const TASK_SITE_SETUP_CHECKLIST = 'home-task-site-setup-checklist';
 export const TASK_WEBINARS = 'home-task-webinars';
 export const TASK_WP_COURSES = 'home-task-wp-courses';
